requirements.txt
Django
python3-nmap
python-whois